Move grunt tools out of config directory
